chore(engine): Cleanup (#1435)
