+++ mod/execution/pkg/client/engine.go
@@ -40,15 +40,18 @@ func (s *EngineClient[ExecutionPayloadT]) NewPayload(
 	versionedHashes []common.ExecutionHash,
 	parentBeaconBlockRoot *primitives.Root,
 ) (*common.ExecutionHash, error) {
-	startTime := time.Now()
-	defer s.metrics.measureNewPayloadDuration(startTime)
-	dctx, cancel := context.WithTimeoutCause(
-		ctx, s.cfg.RPCTimeout, engineerrors.ErrEngineAPITimeout,
+	var (
+		startTime    = time.Now()
+		dctx, cancel = context.WithTimeoutCause(
+			ctx, s.cfg.RPCTimeout, engineerrors.ErrEngineAPITimeout,
+		)
 	)
+
+	defer s.metrics.measureNewPayloadDuration(startTime)
 	defer cancel()
 
 	// Call the appropriate RPC method based on the payload version.
-	result, err := s.callNewPayloadRPC(
+	result, err := s.Eth1Client.NewPayload(
 		dctx,
 		payload,
 		versionedHashes,
@@ -76,28 +79,6 @@ func (s *EngineClient[ExecutionPayloadT]) NewPayload(
 	return processPayloadStatusResult(result)
 }
 
-// callNewPayloadRPC calls the engine_newPayloadVX method via JSON-RPC.
-func (s *EngineClient[ExecutionPayloadT]) callNewPayloadRPC(
-	ctx context.Context,
-	payload ExecutionPayload,
-	versionedHashes []common.ExecutionHash,
-	parentBeaconBlockRoot *primitives.Root,
-) (*engineprimitives.PayloadStatusV1, error) {
-	switch payload.Version() {
-	case version.Deneb:
-		return s.NewPayloadV3(
-			ctx,
-			payload,
-			versionedHashes,
-			parentBeaconBlockRoot,
-		)
-	case version.Electra:
-		return nil, errors.New("TODO: implement Electra payload")
-	default:
-		return nil, engineerrors.ErrInvalidPayloadType
-	}
-}
-
 // ForkchoiceUpdated calls the engine_forkchoiceUpdatedV1 method via JSON-RPC.
 func (s *EngineClient[ExecutionPayloadT]) ForkchoiceUpdated(
 	ctx context.Context,
@@ -105,11 +86,14 @@ func (s *EngineClient[ExecutionPayloadT]) ForkchoiceUpdated(
 	attrs engineprimitives.PayloadAttributer,
 	forkVersion uint32,
 ) (*engineprimitives.PayloadID, *common.ExecutionHash, error) {
-	startTime := time.Now()
-	defer s.metrics.measureForkchoiceUpdateDuration(startTime)
-	dctx, cancel := context.WithTimeoutCause(
-		ctx, s.cfg.RPCTimeout, engineerrors.ErrEngineAPITimeout,
+	var (
+		startTime    = time.Now()
+		dctx, cancel = context.WithTimeoutCause(
+			ctx, s.cfg.RPCTimeout, engineerrors.ErrEngineAPITimeout,
+		)
 	)
+
+	defer s.metrics.measureForkchoiceUpdateDuration(startTime)
 	defer cancel()
 
 	// If the suggested fee recipient is not set, log a warning.
@@ -122,7 +106,12 @@ func (s *EngineClient[ExecutionPayloadT]) ForkchoiceUpdated(
 		)
 	}
 
-	result, err := s.callUpdatedForkchoiceRPC(dctx, state, attrs, forkVersion)
+	result, err := s.Eth1Client.ForkchoiceUpdated(
+		dctx,
+		state,
+		attrs,
+		forkVersion,
+	)
 
 	if err != nil {
 		if errors.Is(err, engineerrors.ErrEngineAPITimeout) {
@@ -140,24 +129,6 @@ func (s *EngineClient[ExecutionPayloadT]) ForkchoiceUpdated(
 	return result.PayloadID, latestValidHash, nil
 }
 
-// updateForkChoiceByVersion calls the engine_forkchoiceUpdatedVX method via
-// JSON-RPC.
-func (s *EngineClient[ExecutionPayloadT]) callUpdatedForkchoiceRPC(
-	ctx context.Context,
-	state *engineprimitives.ForkchoiceStateV1,
-	attrs engineprimitives.PayloadAttributer,
-	forkVersion uint32,
-) (*engineprimitives.ForkchoiceResponseV1, error) {
-	switch forkVersion {
-	case version.Deneb:
-		return s.ForkchoiceUpdatedV3(ctx, state, attrs)
-	case version.Electra:
-		return nil, errors.New("TODO: implement Electra hardfork")
-	default:
-		return nil, engineerrors.ErrInvalidPayloadAttributes
-	}
-}
-
 // GetPayload calls the engine_getPayloadVX method via JSON-RPC. It returns
 // the execution data as well as the blobs bundle.
 func (s *EngineClient[ExecutionPayloadT]) GetPayload(
@@ -165,28 +136,20 @@ func (s *EngineClient[ExecutionPayloadT]) GetPayload(
 	payloadID engineprimitives.PayloadID,
 	forkVersion uint32,
 ) (engineprimitives.BuiltExecutionPayloadEnv[ExecutionPayloadT], error) {
-	startTime := time.Now()
-	defer s.metrics.measureGetPayloadDuration(startTime)
-	dctx, cancel := context.WithTimeoutCause(
-		ctx, s.cfg.RPCTimeout, engineerrors.ErrEngineAPITimeout,
+	var (
+		startTime    = time.Now()
+		dctx, cancel = context.WithTimeoutCause(
+			ctx,
+			s.cfg.RPCTimeout,
+			engineerrors.ErrEngineAPITimeout,
+		)
 	)
+
+	defer s.metrics.measureGetPayloadDuration(startTime)
 	defer cancel()
 
-	// Determine what version we want to call.
-	var fn func(
-		context.Context, engineprimitives.PayloadID,
-	) (engineprimitives.BuiltExecutionPayloadEnv[ExecutionPayloadT], error)
-	switch forkVersion {
-	case version.Deneb:
-		fn = s.GetPayloadV3
-	case version.Electra:
-		return nil, errors.New("TODO: implement Electra getPayload")
-	default:
-		return nil, engineerrors.ErrInvalidGetPayloadVersion
-	}
-
 	// Call and check for errors.
-	result, err := fn(dctx, payloadID)
+	result, err := s.Eth1Client.GetPayload(dctx, payloadID, forkVersion)
 	switch {
 	case err != nil:
 		if errors.Is(err, engineerrors.ErrEngineAPITimeout) {
